+++ src/drivers/warmers/BaseCacheWarmer.php
@@ -56,7 +56,7 @@ abstract class BaseCacheWarmer extends SavableComponent implements CacheWarmerIn
     /**
      * @inheritdoc
      */
-    public function warmSite(int $siteId, callable $setProgressHandler = null, int $delay = null)
+    public function warmSite(int $siteId, callable $setProgressHandler = null, int $delay = null, bool $queue = true)
     {
         // Get custom site URIs for the provided site only
         $groupedSiteUris = SiteUriHelper::getSiteUrisGroupedBySite(Blitz::$plugin->settings->getCustomSiteUris());
@@ -67,13 +67,13 @@ abstract class BaseCacheWarmer extends SavableComponent implements CacheWarmerIn
             $customSiteUris
         );
 
-        $this->warmUris($siteUris, $setProgressHandler, $delay);
+        $this->warmUris($siteUris, $setProgressHandler, $delay, $queue);
     }
 
     /**
      * @inheritdoc
      */
-    public function warmAll(callable $setProgressHandler = null, int $delay = null)
+    public function warmAll(callable $setProgressHandler = null, int $delay = null, bool $queue = true)
     {
         $event = new RefreshCacheEvent();
         $this->trigger(self::EVENT_BEFORE_WARM_ALL_CACHE, $event);
@@ -87,7 +87,7 @@ abstract class BaseCacheWarmer extends SavableComponent implements CacheWarmerIn
             Blitz::$plugin->settings->getCustomSiteUris()
         );
 
-        $this->warmUris($siteUris, $setProgressHandler, $delay);
+        $this->warmUris($siteUris, $setProgressHandler, $delay, $queue);
 
         if ($this->hasEventHandlers(self::EVENT_AFTER_WARM_ALL_CACHE)) {
             $this->trigger(self::EVENT_AFTER_WARM_ALL_CACHE, $event);